+++ src/storage/index.ts
@@ -1,0 +1,8 @@
+export type {
+  HeaderWritableStorageAdapter,
+  SessionSnapshot,
+  StorageAdapter,
+} from "./storage-adapter.js";
+export { canSaveHeader } from "./storage-adapter.js";
+export { FileStorageAdapter } from "./file-storage-adapter.js";
+export { MemoryStorageAdapter } from "./memory-storage-adapter.js";